% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{auth0_api_token}
\alias{auth0_api_token}
\title{Retrieve an Auth0 API Token}
\usage{
auth0_api_token(config_file)
}
\value{
A character string representing the Auth0 API token.
}
\description{
This function retrieves an Auth0 API token by either using an existing token or by making
an API call to obtain a new token. It uses client credentials grant.
}
\details{
@inheritParams auth0_config
}
\examples{
\dontrun{
  client <- "your_client_id"
  secret <- "your_client_secret"
  auth0_api_token(client, secret)
}

}
