% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auth0_find_config_file}
\alias{auth0_find_config_file}
\title{Find the configuration file.}
\usage{
auth0_find_config_file()
}
\value{
Character vector of length one contaning the path of the
\code{_auth0.yml} file.
}
\description{
Tries to find the path to the \code{_auth0.yml} file. First, it tries to get
this info from \code{options(auth0_config_file = )}. If this option is \code{NULL}
(the default) it tries to find the \code{_auth0.yml} within the working
directory. If the file does not exist, it raises an error.
}
\seealso{
\code{\link{use_auth0}}.
}
