% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{auth0_user_management}
\alias{auth0_user_management}
\title{Manage Auth0 Users}
\usage{
auth0_user_management(user_id, action = "get", body = list())
}
\arguments{
\item{user_id}{Character string, the ID of the user you want to manage.}

\item{action}{Character string, the action to be performed; either 'get' or 'update'.
Default is 'get'.}

\item{body}{List, additional body parameters for the 'update' action.}
}
\value{
A list representing the parsed response from the Auth0 API.
}
\description{
This function provides an interface for managing users in Auth0. It supports
actions for getting and updating user details.
}
\examples{
\dontrun{
  user_id <- "user_auth0_id"
  action <- "get"
  auth0_user_management(user_id, action)

  user_id <- "user_auth0_id"
  action <- "update"
  body <- list(user_metadata = list(theme='dark'))
  auth0_user_management(user_id, action, body)
}

}
