% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logout}
\alias{logout}
\title{Log out of an auth0 app}
\usage{
logout(config_file = NULL)
}
\arguments{
\item{config_file}{Path to YAML configuration file.}
}
\description{
Log the current user out of an auth0 shiny app.
}
\details{
If you want to use a different configuration file you can set the \code{auth0_config_file}
option with: \code{options(auth0_config_file = "path/to-file")}.
}
\seealso{
\code{\link[auth0]{logoutButton}},
\code{\link[auth0]{logout_url}}
}
