% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logout_url}
\alias{logout_url}
\title{Generate a logout URL}
\usage{
logout_url(config_file = NULL)
}
\arguments{
\item{config_file}{Path to YAML configuration file.}
}
\value{
URL string to log out.
}
\description{
Generate a URL that will log the user out of the app if visited.
}
\details{
If you want to use a different configuration file you can set the \code{auth0_config_file}
option with: \code{options(auth0_config_file = "path/to-file")}.
}
\seealso{
\code{\link[auth0]{logoutButton}},
\code{\link[auth0]{logout}}
}
