% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shinyAuth0App}
\alias{shinyAuth0App}
\title{Create a Shiny app object with Auth0 Authentication}
\usage{
shinyAuth0App(ui, server)
}
\arguments{
\item{ui}{an ordinary UI object to create shiny apps.}

\item{server}{an ordinary server object to create shiny apps.}
}
\description{
This function modifies ui and server objects to run using Auth0
authentication.
}
\details{
If you want to use a diferent configuration file you can set the \code{auth0_config_file}
option with: \code{options(auth0_config_file = "path/to-file")}.
}
