% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{auth0_user_management}
\alias{auth0_user_management}
\title{Manage Auth0 Users}
\usage{
auth0_user_management(
  user_id,
  action = "get",
  body = list(),
  config_file = NULL
)
}
\arguments{
\item{user_id}{Character string, the ID of the user you want to manage.}

\item{action}{Character string, the action to be performed; either 'get' or 'update'.
Default is 'get'.}

\item{body}{List, additional body parameters for the 'update' action.}
}
\value{
A list representing the parsed response from the Auth0 API.
}
\description{
This function provides an interface for managing users in Auth0. It supports
actions for getting and updating user details.
}
