% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shinyAuth0App}
\alias{shinyAuth0App}
\title{Create a Shiny app object with Auth0 Authentication}
\usage{
shinyAuth0App(ui, server, config_file = "./_auth0.yml")
}
\arguments{
\item{ui}{an ordinary UI object to create shiny apps.}

\item{server}{an ordinary server object to create shiny apps.}

\item{config_file}{Path to YAML configuration file.}
}
\description{
This function modifies ui and server objects to run using Auth0
authentication.
}
