% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shinyAuth0App}
\alias{shinyAuth0App}
\title{Create a Shiny app object with Auth0 Authentication}
\usage{
shinyAuth0App(ui, server, config_file = NULL)
}
\arguments{
\item{ui}{an ordinary UI object to create shiny apps.}

\item{server}{an ordinary server object to create shiny apps.}

\item{config_file}{path to YAML configuration file.}
}
\description{
As of auth0 0.1.2, \code{shinAuth0App()} has
been renamed to \code{\link[=shinyAppAuth0]{shinyAppAuth0()}} for consistency.
}
